Reorganizing code structure
Clear folders and filenames for all protocols and structs
